+++ Makie/src/basic_recipes/datashader.jl
@@ -175,15 +175,15 @@ module Aggregation
                 out[i, j] = update(op, out[i, j], z)
             end
         end
-
         mini, maxi = Inf, -Inf
-        map!(pixelbuffer, aggbuffer) do x
-            final_value = value(op, x)
+
+        for i in eachindex(pixelbuffer, aggbuffer)
+            final_value = value(op, aggbuffer[i])
             if isfinite(final_value)
                 mini = min(final_value, mini)
                 maxi = max(final_value, maxi)
             end
-            return final_value
+            pixelbuffer[i] = final_value
         end
         c.data_extrema = (mini, maxi)
         return c